file rename: appium-service-configurator-factory.ts => appium-adb-wrapper.ts
