+++ src/electron/platform/android/appium-adb-wrapper-factory.ts
